Readded external (wtf), added bin2c to convert image and data to array
to be usable as RWops in the engine. Tryed to add physfs but it crash on windows. Edited enum for log to ICE_LOGTYPE_ in the name and added macro to the ICE_Log to write ICE_Log_Error() and not ICE_Log(ICE_LOGTYPE_ERROR, format_, ...).// Ressources folder in engine for "binary" array, it will replace the "raw" folder.