Edit in thread screenshot (bug when mass screenshot at one time)
added macrotools with STRINGIZE who transforme any text to C string. Own file for screenshot function in graphics. Made everything in texture load "RW" and made a wrapper. Use res:// to access physfs loaded zip with. Macro to load from memory with only the array token to write and with logs. Function to check the res:// from a string and Macrowrapper + function to load zip from memory (so combo bin2c and physfs).